+++ typescript/@ragbits/api-client/src/autogen.types.ts
@@ -56,6 +56,15 @@ export const MessageRole = {
 export type MessageRole = TypeFrom<typeof MessageRole>
 
 /**
+ * Represents the AuthType enum
+ */
+export const AuthType = {
+    Credentials: 'credentials',
+} as const
+
+export type AuthType = TypeFrom<typeof AuthType>
+
+/**
  * Represents the context of a chat conversation.
  */
 export interface ChatContext {
@@ -64,6 +73,7 @@ export interface ChatContext {
     state: {
         [k: string]: unknown
     }
+    session_id: string | null
     [k: string]: unknown
 }
 
@@ -219,6 +229,7 @@ export interface ConfigResponse {
      * Debug mode flag
      */
     conversation_history: boolean
+    authentication: AuthenticationConfig
 }
 
 /**
@@ -267,6 +278,105 @@ export interface FeedbackRequest {
      * Additional feedback details
      */
     payload: {
+        [k: string]: unknown
+    }
+}
+
+/**
+ * Configuration for authentication.
+ */
+export interface AuthenticationConfig {
+    /**
+     * Enable/disable authentication
+     */
+    enabled: boolean
+    /**
+     * List of available authentication types
+     */
+    auth_types: AuthType[]
+}
+
+/**
+ * Request body for user login
+ */
+export interface CredentialsLoginRequest {
+    /**
+     * Username
+     */
+    username: string
+    /**
+     * Password
+     */
+    password: string
+}
+
+/**
+ * Represents a JWT authentication jwt_token.
+ */
+export interface JWTToken {
+    access_token: string
+    token_type: string
+    expires_in: number
+    refresh_token: string | null
+    user: User
+}
+
+/**
+ * Request body for user login
+ */
+export interface LoginRequest {
+    /**
+     * Username
+     */
+    username: string
+    /**
+     * Password
+     */
+    password: string
+}
+
+/**
+ * Response body for successful login
+ */
+export interface LoginResponse {
+    /**
+     * Whether login was successful
+     */
+    success: boolean
+    /**
+     * User information
+     */
+    user: User | null
+    /**
+     * Error message if login failed
+     */
+    error_message: string | null
+    /**
+     * Access jwt_token
+     */
+    jwt_token: JWTToken | null
+}
+
+/**
+ * Request body for user logout
+ */
+export interface LogoutRequest {
+    /**
+     * Session ID to logout
+     */
+    token: string
+}
+
+/**
+ * Represents an authenticated user.
+ */
+export interface User {
+    user_id: string
+    username: string
+    email: string | null
+    full_name: string | null
+    roles: string[]
+    metadata: {
         [k: string]: unknown
     }
 }